[UPDATE] divide signup page into components
