+++ package.json
@@ -30,9 +30,7 @@
     "react-icons": "^4.3.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
-    "reactjs-line-login": "^1.2.3",
     "sass": "^1.49.7",
-    "universal-cookie": "^4.0.4",
     "web-vitals": "^2.1.4",
     "yup": "^0.32.11"
   },
